[feature/MOB-330] - 1- Make "scaleFiat" method static in CurrencyFormatUtils; 2- Change "formatCurrency" method argument from double to BigDecimal; 3- Create static methods "mapToWalletCurrency" in WalletCurrency.

+++ app/src/main/java/com/asfoundation/wallet/billing/adyen/AdyenPaymentFragment.kt
@@ -182,7 +182,7 @@ class AdyenPaymentFragment : DaggerFragment(), AdyenPaymentView {
       e.printStackTrace()
     }
     app_sku_description?.text = arguments!!.getString(IabActivity.PRODUCT_NAME)
-    val appcValue = formatter.formatCurrency(appcAmount.toDouble(), WalletCurrency.APPCOINS)
+    val appcValue = formatter.formatCurrency(appcAmount, WalletCurrency.APPCOINS)
     appc_price.text = appcValue.plus(" " + WalletCurrency.APPCOINS.symbol)
   }
